scripts/requirements.txt
absl-py==2.3.0
accelerate==1.8.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
bangla==0.0.5
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dateparser==1.1.8
decorator==5.2.1
diffusers==0.33.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
g2pkk==0.1.2
gradio==5.34.2
gradio_client==1.10.3
groovy==0.1.2
grpcio==1.73.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gTTS==2.5.4
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
inflect==7.5.0
ipython==8.37.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.1
jsonlines==1.2.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
murmurhash==1.0.13
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyMatting==1.1.14
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.54.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.11.4
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.51.3
trimesh==4.6.12
triton==3.3.1
TTS==0.22.0
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.34.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
